pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pharma-job-search"
version = "1.0.0"
description = "CLI + dashboard that aggregates pharma/biotech job listings from 5 sources with AI-powered evaluation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "BioTechNerd-Apache" },
]
keywords = ["pharma", "biotech", "job", "search", "cli", "streamlit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]
dependencies = [
    "python-jobspy>=1.1.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "streamlit>=1.30.0",
    "streamlit-aggrid>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
]

[project.scripts]
pharma-job-search = "job_search:main"

[project.urls]
Homepage = "https://github.com/BioTechNerd-Apache/pharma-job-search"
Repository = "https://github.com/BioTechNerd-Apache/pharma-job-search"
Issues = "https://github.com/BioTechNerd-Apache/pharma-job-search/issues"

[tool.setuptools]
py-modules = ["job_search"]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
